+++ CHANGES.md
@@ -4,6 +4,10 @@
   `xcube.core.store.fs.impl.rasterio` when reading GeoTIFF and JPEG2000 files.
   This enables efficient, storage-aware data access without forcing explicit
   rechunking.
+* Added the optional argument `band_as_variable`, which allows to preserve the
+  original dataset structure as returned by `rioxarray`, rather than splitting
+  raster bands into separate data variables. This improves data access patterns and
+  avoids unnecessary transformations. Defaults are set to `False`.
 
 
 ## Changes in 1.13.0